% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confIntICC.R
\name{Aalpha}
\alias{Aalpha}
\title{A(alpha)}
\usage{
Aalpha(alpha, n, d, MSs, MSe)
}
\arguments{
\item{alpha}{See indicated reference.}

\item{n}{See indicated reference.}

\item{d}{See indicated reference.}

\item{MSs}{See indicated reference.}

\item{MSe}{See indicated reference.}
}
\description{
The A(alpha) function from Roussen et al (2003), p. 621.
Used in \code{\link{confIntICC}}.
}
\seealso{
confIntICC
}
