% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behrensTest.R
\name{behrensTest}
\alias{behrensTest}
\title{Behrens version of the t-test for unequal variances}
\usage{
behrensTest(
  x,
  y,
  conf.level = 0.95,
  alternative = c("two.sided", "greater", "less")
)
}
\arguments{
\item{x}{Numeric vector of data values.}

\item{y}{Numeric vector of data values.}

\item{conf.level}{Confidence level of the interval, default is 0.95.}

\item{alternative}{Character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater", and "less".}
}
\description{
Behrens version of the t-test for unequal variances.
}
\details{
Computes behrens version of the t-test for unequal variances
based on approximate solution as in Box & Tiao, 1973, Section 2.5.3;
see also Armitage, Berry, Matthews, 2002, Section 4.3.
}
\examples{
weightWomen <- c(38.9, 61.2, 73.3, 21.8, 63.4, 64.6, 48.4, 48.8, 48.5)
weightMen <-   c(67.8, 60  , 63.4, 76,   89.4, 73.3, 67.3, 61.3)

behrensTest(x = weightWomen, y = weightMen)
behrensTest(x = weightWomen, y = weightMen, alternative="less")
}
