% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confIntCorrelation.R
\name{confIntCorrelation}
\alias{confIntCorrelation}
\title{Confidence interval for correlation coefficient using Fisher's
transformation}
\usage{
confIntCorrelation(
  x,
  y,
  conf.level = 0.95,
  method = c("spearman", "pearson"),
  type = c("t", "z")
)
}
\arguments{
\item{x}{Vector containing the first variable.}

\item{y}{Vector of same length as \code{x} containing the second variable.}

\item{conf.level}{Confidence level for confidence interval. Default is 0.95.}

\item{method}{Correlation coefficient to be used: "spearman" (default) or "pearson".}

\item{type}{Quantile to be used: "t" (default) or "z".}
}
\value{
List with entries:
\item{estimate}{Value of correlation coefficient.}
\item{ci}{Computed confidence interval.}
\item{p.value}{\eqn{p}-value for a test on \eqn{\rho = 0} based on the transformation.}
\item{n}{Number of observations.}
\item{p2}{\eqn{p}-value based on the \code{R} function \code{cor.est}.}
}
\description{
Compute a confidence interval for a correlation coefficient \eqn{r} using
the variance-stabilizing transformation
\deqn{z = \tanh^{-1}(r) = 0.5 \log((1 + r) / (1 - r)),}
known as Fisher's \eqn{z}-transformation.
Independent of the true correlation \eqn{\rho}, \eqn{z} is approximately
normally distributed with variance \eqn{(n-3)^{-1}}. This enables the construction
of a Wald-type confidence interval. Back-transformating this interval yields a confidence
interval for \eqn{r}. An advantage of this method is
that the confidence interval is contained in \eqn{(-1, 1)}.
}
\examples{

n <- 40
x <- runif(n = n)
y <- 2 * x + 0.5 * rnorm(n = n)
confIntCorrelation(x = x, y = y)

}
\author{
Kaspar Rufibach \cr \email{kaspar.rufibach@gmail.com}
}
\keyword{htest}
