% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confIntKM.R
\name{confIntKM}
\alias{confIntKM}
\alias{confIntKM_t0}
\title{Confidence intervals for a survival curve at a fixed time point}
\usage{
confIntKM(time, event, t0, conf.level = 0.95)
}
\arguments{
\item{time}{Event times, censored or observed.}

\item{event}{Censoring indicator, 1 for event, 0 for censored.}

\item{t0}{Numeric vector of time points to compute confidence interval for.}

\item{conf.level}{Confidence level for confidence interval.}
}
\value{
A matrix with the following columns:
\item{t0}{Time points.}
\item{S_t0}{Value of survival curve at \code{t0}.}
\item{lower.ci}{Lower limits of confidence interval(s).}
\item{upper.ci}{Upper limits of confidence interval(s).}
}
\description{
Computes Wald confidence intervals for a Kaplan-Meier survival curve at
a fixed time point. The variance is computed according to Peto's formula and the
confidence interval is computed using a logit-transformation, to ensure that
it lies in \eqn{(0, 1)}. Alternatives are given in the examples
below.
}
\note{
The variance according to Peto's formula tends to be more conservative
than that based on Greenwood's formula.
}
\examples{

## use Acute Myelogenous Leukemia survival data contained in package 'survival'
time <- leukemia[, 1]
event <- leukemia[, 2]
formula <- Surv(time = time, event = event) ~ 1
plot(survfit(formula = formula, conf.type = "none"), mark = "/", col = 1:2)
confIntKM(time = time, event = event, t0 = c(10, 25, 50), conf.level = 0.95)


## an alternative is the log-log confidence interval using Greenwood's
## variance estimate
t0 <- 10
fit <- survfit(formula = formula, conf.int = 0.95, conf.type = "log-log", 
               type = "kaplan", error = "greenwood")
dat <- cbind(fit$time, fit$surv, fit$lower, fit$upper)
dat <- dat[dat[, 1] >= t0, ]
dat[1, 3:4]


## this same confidence interval can also be computed using the 
## package km.ci
if(require(km.ci)) {
    ci.km <- km.ci(survfit(formula = formula), conf.level = 0.95,
                   method = "loglog")
    dat.km <- cbind(ci.km$time, ci.km$surv, ci.km$lower, ci.km$upper)
    dat.km <- dat.km[dat.km[, 1] >= t0, 3:4]
    dat.km[1, ]
}

}
\author{
Kaspar Rufibach \cr \email{kaspar.rufibach@gmail.com}
}
\keyword{htest}
\keyword{survival}
