% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confIntRateRatio.R
\name{confIntRateRatio}
\alias{confIntRateRatio}
\title{Compute confidence interval for a rate ratio}
\usage{
confIntRateRatio(x, t, conf.level = 0.95)
}
\arguments{
\item{x}{A vector of length 2 containing the number of events in each group.}

\item{t}{A numeric vector of length 2 containing the total observation time
in each group.}

\item{conf.level}{A numeric vector of length 1 containing the confidence
level for the confidence interval.}
}
\value{
A named numeric vector of length 3 containing the rate ratio and
    the limits of the confidence interval.
}
\description{
A method to compute a confidence interval for a rate ratio
    is provided. The method is based on a Wald interval for the log rate
    ratio.
}
\examples{
    x <- c(30, 50)
    t <- c(100, 120)
    confIntRateRatio(x, t)

}
\references{
Held, L., Rufibach, K. and Seifert, B. (2013)
    \emph{Medizinische Statistik - Konzepte, Methoden, Anwendungen}.
    Section 8.3.
}
\seealso{
\code{link[biostatUZH]{confIntRateDiff}}
}
\author{
Leonhard Held
}
\keyword{htest}
