% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confIntSquareAdd.R
\name{confIntSquareAdd}
\alias{confIntSquareAdd}
\title{Square-and-Add Method for Confidence Interval of Difference in Point Estimates}
\usage{
confIntSquareAdd(theta1, lower1, upper1, theta2, lower2, upper2)
}
\arguments{
\item{theta1}{Numeric value of the point estimate for group 1.}

\item{lower1}{Numeric value of the lower bound of the confidence interval for group 1.}

\item{upper1}{Numeric value of the upper bound of the confidence interval for group 1.}

\item{theta2}{Numeric value of the point estimate for group 2.}

\item{lower2}{Numeric value of the lower bound of the confidence interval for group 2.}

\item{upper2}{Numeric value of the upper bound of the confidence interval for group 2.}
}
\value{
A list with the entries: \itemize{ \item difference: Estimated difference (group 1 - group 2).
\item CI: Lower and upper bound of the confidence interval for the difference.}
}
\description{
This function computes the confidence interval for the difference between 
two point estimates using the square-and-add method.
}
\details{
The function returns a confidence interval of the same level 
as the confidence intervals for theta1 and theta2
}
\examples{
ci1 <- wilson(x=11, n=12)
ci2 <- wilson(x=1, n=12)
confIntSquareAdd(theta1 = ci1[2], lower1=ci1[1], upper1=ci1[3],
                 theta2 = ci2[2], lower2=ci2[1], upper2=ci2[3]) 
}
\references{
Newcombe, R.G. (1998). Interval estimation for the difference 
between independent proportions: comparison of eleven methods, 
\emph{Statistics in Medicine}, \bold{17}, 873--890.

Newcombe, R.G. (2013). \emph{Confidence Intervals for Proportions 
and Related Measures of Effect Size}, CRC Press, 
2nd Edition, Chapter 7.3.
}
\author{
Manuel Pfister, Leonhard Held, Charlotte Micheloud
}
