% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faganPlot.R
\name{faganPlot}
\alias{faganPlot}
\alias{faganLine}
\title{Fagan-Nomogram}
\usage{
faganPlot(
  probPreInit = c(0.1, 0.2, 0.5, 1, 2, 5, 10, 20, 30, 40, 50, 70, 85),
  text = 0.7,
  language = c("german", "english"),
  title = "Fagan - Nomogramm"
)
}
\arguments{
\item{probPreInit}{Pre-test probabilities to be used for the vertical line
to the left. May be a number or a vector withe entries in \eqn{(0,1)}.}

\item{text}{Size of the text displayed in the plot.}

\item{language}{Choose language.}

\item{title}{Title to be added to plot, or \code{NA} if to be left blank.}
}
\value{
Nothing is returned. \code{faganPlot} draws a Fagan-Nomogram and
\code{faganLine} adds lines for specific pre-test probabilities and
likelihood ratios to the plot.
}
\description{
Fagan-Nomogram
}
\examples{

# empty Fagan plot
faganPlot()

# Fagan lines
faganLine(prob.pre = 0.75, lik.ratio = c(0.9/0.37, 0.1/0.63))

}
\references{
Fagan, T.J. (1975). Letter: Nomogram for Bayes Theorem. \emph{N.
Engl. J. Med}, \bold{293}, 257.
}
\author{
Kaspar Rufibach \cr \email{kaspar.rufibach@gmail.com}
}
\keyword{aplot}
\keyword{dplot}
