% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gailSimon.R
\name{gailSimon}
\alias{gailSimon}
\title{Test for qualitative interaction by Gail and Simon}
\usage{
gailSimon(thetahat, se)
}
\arguments{
\item{thetahat}{A numeric vector of length > 1. Denotes the effect estimates
in the different subgroups.}

\item{se}{A numeric vector of the same length as \code{thetahat}. Contains 
the standard errors of the corresponding effect estimates in each subgroup.}
}
\value{
A named numeric vector of length 2 containing the p-value and the 
corresponding test statistic according to the Gail-Simon test.
}
\description{
This function computes the p-value of the Gail-Simon test for
qualitative interaction from effect estimates and the corresponding 
standard errors in a number of subgroups
}
\examples{
rd <- c(0.163, -0.114, -0.047, -0.151)
rd.se <- c(0.0788, 0.0689, 0.0614, 0.0547)
## Gail and Simon test
gailSimon(thetahat = rd, se = rd.se)
}
\references{
Gail, M., & Simon, R. (1985). Testing for Qualitative Interactions between 
Treatment Effects and Patient Subsets. \emph{Biometrics, 41}(2), 361–372. 
\url{https://doi.org/10.2307/2530862}
}
\author{
Leonhard Held
}
