% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.wholenumber.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Check if (close to) whole number}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{Numeric vector}

\item{tol}{Numeric vector of length 1 specifying the tolerance.
Default is \code{.Machine$double.eps^0.5}.}
}
\description{
Check what elements of a vector are (close to) whole numbers by
\code{abs(x - round(x)) < tol}.
}
\examples{
is.wholenumber(1L:5L)
is.wholenumber(1:5)
is.wholenumber(1:5 + .Machine$double.eps^0.5 / 2)
is.wholenumber(1:5 + .Machine$double.eps^0.5)
}
