% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioStatUZH-package.R
\docType{data}
\name{larynx}
\alias{larynx}
\title{Survival Times of Larynx Cancer Patients}
\format{
A data frame with 90 observations on the following 5 variables.
\describe{
\item{stage}{Disease stage (1-4) from TNM cancer staging classification.}
\item{time}{Time from first treatment until death, or end of study.}
\item{age}{Age at diagnosis.}
\item{year}{Year of diagnosis.}
\item{death}{Indicator of death [1, if patient died at time t; 0, otherwise].} }
}
\description{
A study of 90 males with laryngeal cancer was performed, comparing survival
times. Each patient's age, year of diagnosis, and disease stage was noted.
Study published in Kardaun (1983), and available at the website for Klein
and Moeschberger (2003).
}
\examples{

library(survival)
data(larynx)
Surv(larynx$time, larynx$death)

}
\references{
Kardaun, O. (1983). Statistical survival analysis of male
larynx-cancer patients-a case study.  \emph{Statistica Neerlandica},
\bold{37}, 103--125.

Klein, J. and Moeschberger, M. (2003). \emph{Survival analysis: techniques
for censored and truncated data}.  Springer.
}
\keyword{datasets}
