% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\alias{ilogit}
\alias{expit}
\title{Logit and inverse-logit function}
\usage{
logit(x)
}
\description{
Compute logit and inverse logit function.
}
\details{
Used by \code{\link{faganPlot}}, \code{\link{faganLine}}.
}
\author{
Kaspar Rufibach \cr \email{kaspar.rufibach@gmail.com}
}
