% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantelHaenszel.R
\name{mantelHaenszel}
\alias{mantelHaenszel}
\title{Cochran-Mantel-Haenszel Chi-Squared Test for Count Data}
\usage{
mantelHaenszel(exposure, outcome, stratum)
}
\arguments{
\item{exposure}{Binary variable coding whether patient was exposed (1) or
not (0).}

\item{outcome}{Binary variable coding outcome of patient.}

\item{stratum}{Factor object with at least 2 levels identifying to which
stratum the corresponding elements in \code{exposure} and \code{outcome} belong.}
}
\value{
\item{tab}{Table that counts numbers of strata for each case-control
combination.}
\item{test.stat}{Test statistic for \eqn{chi^2} test.}
\item{p.val}{\eqn{p}-value of \eqn{chi^2} test.}
}
\description{
Compute Cochran-Mantel-Haenszel chi-squared test of the null that two
nominal variables are conditionally independent in each stratum, assuming
that there is no three-way interaction.
}
\examples{

# generate data
set.seed(1977)
data <- data.frame(exposure = rep(c(1, 0, 0, 0, 0), 41),
                   outcome = sample(x = c(rep(1, 62), rep(0, 5 * 41 - 62))),
                   stratum = rep(1:41, each = 5))

# via conditional logistic regression
logreg <- clogit(outcome ~ exposure + strata(stratum), method = "approximate",
                 data = data)
summary(logreg)

# R function in library 'stats'
mh <- with(data, mantelhaen.test(x = outcome, y = exposure, z = stratum))

# this function
mH <- with(data, mantelHaenszel(exposure = exposure,
                                outcome = outcome,
                                stratum = stratum))

# compare p-values
summary(logreg)$coef[5]
mh$p.value
mH$p.val

}
\references{
Agresti, A. (2002). \emph{Categorical data analysis}. Wiley, New York,
}
\seealso{
Similar functionality is provided in \code{\link[stats]{mantelhaen.test}}
and \code{\link[survival]{clogit}}. See the examples below for a
comparison.
}
\author{
Kaspar Rufibach \cr \email{kaspar.rufibach@gmail.com}
}
\keyword{htest}
\keyword{survival}
