% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationSamplePlot.R
\name{populationSamplePlot}
\alias{populationSamplePlot}
\title{Plot that shows population and sample drawn from it}
\usage{
populationSamplePlot(language = c("german", "english")[1])
}
\arguments{
\item{language}{Language of the text pieces added to the plot.}
}
\value{
Nothing is ouutput, only a plot is generated,
}
\description{
We frequently use this plot in presentations and teaching, to explain the
connection between a population and a sample drawn from it.
}
\author{
Kaspar Rufibach \cr \email{kaspar.rufibach@gmail.com}
}
\keyword{aplot}
\keyword{dplot}
