% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerZtest.R
\name{power.z.test}
\alias{power.z.test}
\title{Power calculations for two sample z tests}
\usage{
power.z.test(
  n = NULL,
  delta = NULL,
  sd = 1,
  sig.level = 0.05,
  power = NULL,
  type = c("two.sample", "one.sample"),
  alternative = c("two.sided", "one.sided")
)
}
\arguments{
\item{n}{Number of observations (per group)}

\item{delta}{True difference in means}

\item{sd}{Standard deviation}

\item{sig.level}{Significance level (Type I error probability)}

\item{power}{Power of test (1 minus Type II error probability)}

\item{type}{String specifying the type of t test. Can be abbreviated.}

\item{alternative}{One- or two-sided test. Can be abbreviated.}
}
\value{
Object of class "power.htest", a list of the arguments 
(including the computed one) augmented with method and note elements.
}
\description{
Power calculations for two sample z tests
}
\details{
Exactly one of the parameters \code{n}, \code{delta}, 
\code{power}, \code{sd}, and \code{sig.level} must be passed as \code{NULL}, 
and that parameter is determined from the others. Notice that the last two 
have non-\code{NULL} defaults, so \code{NULL} must be explicitly passed if 
you want to compute them.
}
\examples{

# Calculate sample size
delta = 0.25
sd = 0.4
sig.level = 0.01
power = 0.95
power.z.test(delta = delta, sd = sd, sig.level = sig.level, power = power)

# Calculate the effect size
n = 92
power.z.test(power = power, sd = sd, sig.level = sig.level, n = n)

# Calculate the standard deviation
power.z.test(power = power, delta = delta, sig.level = sig.level, n = n,
             sd = NULL)

# Calculate the type I error
power.z.test(power = power, delta = delta, sig.level = NULL, n = n,
             sd = sd)


# Calculate power
power.z.test(delta = delta, sd = sd, sig.level = sig.level, n = n)

}
