% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printWaldCI.R
\name{printWaldCI}
\alias{printWaldCI}
\title{Confidence Wald confidence interval for a given effect}
\usage{
printWaldCI(theta, se.theta, conf.level = 0.95, FUN = identity, digits = 3)
}
\arguments{
\item{theta}{the effect size.}

\item{se.theta}{the standard error.}

\item{conf.level}{Confidence level. Default is 95\%.}

\item{FUN}{Transformation to applied to effect size and confidence interval.
Default is the identity function.}

\item{digits}{number of digits for the effect.}
}
\value{
Prints and invisiblie returns the effect, the Wald confidence interval, and the p-value.
}
\description{
Takes an effect and its standard error to calculate the Wald
confidence interval and the p-value.
}
\examples{

printWaldCI(theta = 1.7, se.theta = 2.1)
printWaldCI(theta = 1.7, se.theta = 2.1, FUN = exp)

}
\author{
Leonhard Held
}
