% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survReport.R
\name{quantileKM}
\alias{quantileKM}
\title{Quantile of a Kaplan-Meier estimate of a survival function}
\usage{
quantileKM(
  time,
  event,
  group = NULL,
  quantile = 0.5,
  conf.level = 0.95,
  conf.type = c("log-log", "log", "plain", "none")
)
}
\arguments{
\item{time}{Event times, censored or observed.}

\item{event}{Censoring indicator, 1 for event, 0 for censored.}

\item{group}{Indicates groups to compute individual quantile for each group.
Default is \code{NULL}.}

\item{quantile}{Quantile to be computed. Real number in \eqn{[0, 1]}.}

\item{conf.level}{Significance level for confidence interval for the time to
event quantile.}

\item{conf.type}{Type of confidence interval to be computed. For possible
choices see above, and for specifications regarding the different options
see \code{\link[survival]{survfit}}.}
}
\value{
A list containing a matrix with columns:
\item{n}{Number of observations used.}
\item{events}{Number of events.}
\item{quantile}{Quantile estimate.}
\item{lower.ci}{Lower limit of confidence interval.}
\item{upper.ci}{Upper limit of confidence interval.}\cr
and the p-value of the test for the difference between the survival curves;
see \code{\link[survival]{survdiff}} for more information.
}
\description{
In survival analysis, often quantiles (e.g., the median) of an
estimated survival function are of interest. The \code{\link[survival]{survfit}}
function in the  'survival' packages of version \eqn{< 2.35} computes median time to event.
However, the output of that function is such that the median is not
conveniently accessible by the user. This function
makes the median and any other quantile accessible by the user.
}
\note{
The function is based on the function \code{\link[survival]{survfit}}.
}
\examples{

## use Acute Myelogenous Leukemia survival data contained in package 'survival'
time <- leukemia[, 1]
event <- leukemia[, 2]
group <- as.factor(leukemia[, 3])
plot(survfit(Surv(time, event) ~ group, conf.type = "none"), mark = "/", col = 1:2)

## median time to event
quantileKM(time, event, group = group, quantile = 0.5, conf.level = 0.95, 
           conf.type = "log")
 
## comparison to standard function (median time to event not accessible by user)
survfit(Surv(time, event) ~ group, conf.int = 0.95)

## compute 0.25 quantile
quantileKM(time = time, event = event, group = group, quantile = 0.25,
           conf.type = "log")

}
\references{
Confidence intervals are computed according to

Brookmeyer, R. and Crowley, J. (1982).  A Confidence Interval for the Median
Survival Time.  \emph{Biometrics}, \bold{38}, 29--41.
}
\seealso{
\code{\link[survival]{survfit}}, \code{\link[survival]{survdiff}}
}
\author{
Kaspar Rufibach
}
\keyword{htest}
\keyword{survival}
