% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardErrorAUC.R
\name{standardErrorAUC}
\alias{standardErrorAUC}
\title{Standard Error of AUC}
\usage{
standardErrorAUC(cases, controls)
}
\arguments{
\item{cases}{Values of the continuous variable for the cases.}

\item{controls}{Values of the continuous variable for the controls.}
}
\value{
The standard error.
}
\description{
Computes the standard error of the area under the curve.
}
\references{
The computation follows Chapter 10 in

Altman, D.G., Machin, D., Bryant, T.N. and Gardner, M.J. (2001).
\emph{Statistics with confidence}. 2nd Edition, 2000. BMJ Books.
}
\author{
Leonhard Held
}
\keyword{htest}
\keyword{univar}
