% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confIntAUC.R
\name{summaryROC}
\alias{summaryROC}
\title{ROC curve and an asymptotic confidence interval for AUC}
\usage{
summaryROC(cases, controls, conf.level = 0.95)
}
\arguments{
\item{cases}{Values of the continuous variable for the cases.}

\item{controls}{Values of the continuous variable for the controls.}

\item{conf.level}{Confidence level of confidence interval.}
}
\value{
\item{x.val}{1-specificity of the test, so the values on the \eqn{x}-axis
of a ROC plot.}
\item{y.val}{Sensitivity of the test, so the values on the \eqn{y}-axis of
a ROC plot.}
\item{ppvs}{Positive predictive values for each cutoff.}
\item{npvs}{Negative predictive values for each cutoff.}
\item{cutoffs}{Cutoffs used (basically the pooled marker values of
cases and controls).}
\item{res.mat}{Collects the above quantities in a matrix, including
Wilson confidence intervals, computed at at confidence level \code{conf.level}.}
\item{auc}{Area under the ROC curve. This is equal to the value of the
Mann-Whitney test statistic.}
\item{auc.var}{Variance of AUC.}
\item{auc.var.norm}{Variance of AUC if data is assumed to come from a
bivariate normal distribution.}
\item{lowCI}{Lower limit of Wald confidence interval.}
\item{upCI}{Upper limit of Wald confidence interval.}
\item{logitLowCI}{Lower limit of a Wald confidence interval received on
logit scale.}
\item{logitUpCI}{Upper limit of a Wald confidence interval received on
logit scale.}
}
\description{
This function computes the ROC curve for values of a continuous variable
of a group of cases and a group of controls the. Additionally, the AUC
with an asymptotic confidence interval is provided.
}
\note{
The confidence intervals are only valid if observations are
\emph{independent}.
}
\examples{

## simulate data
## --------------
set.seed(1977)
controls <- rnorm(n = 50)
cases <-  rnorm(n = 40, mean = 0.5, sd = 1.5)

## summary of ROC curve
## --------------
res <- summaryROC(cases, controls, conf.level = 0.95)


## alternative bootstrap CI for AUC
## --------------
resBinormBoot <- confIntAUCbinorm(cases = cases, controls = controls, 
                                  conf.level = 0.95, replicates = 1000,
                                  grid = 100)

## alternative bootstrap CI for AUC 
## --------------
resBinormWald <- confIntAUCbinorm(cases, controls, conf.level = 0.95,
                                  method = "wald")

## display results
## --------------
str(res)
str(resBinormBoot)
str(resBinormWald)

## plot ROC curve
## --------------
plot(0, 0, xlim = c(0, 1), ylim = c(0, 1), type = 'l', 
     xlab = "1 - specificity", ylab = "sensitivity", pty = 's')
segments(0, 0, 1, 1, lty = 2)
lines(res$x.val, res$y.val, type = 'l', col = 2, lwd = 2, lty = 2)
lines(resBinormBoot$x.val, resBinormBoot$y.val, type = 'l',
      col = 4, lwd = 2, lty = 2)

}
\references{
The original reference for the computation of the confidence interval is:

Hanley, J.A. and McNeil, B.J. (1982). The meaning and use of the area under
the curve. \emph{Radiology}, \bold{143}, 29--36.

See also:

Pepe, M.S. (2003) \emph{The statistical evaluation of medical tests for
classification and prediction}. Oxford University Press.
}
\seealso{
\code{\link{confIntAUCbinorm}}. Similar functionality is provided in the package \pkg{ROCR}.
}
\author{
Kaspar Rufibach \email{kaspar.rufibach@gmail.com} and Andrea Riebler.
}
\keyword{htest}
