% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survEvents.R
\name{survEvents}
\alias{survEvents}
\title{Compute number of events or power for a survival endpoint}
\usage{
survEvents(
  HR,
  sig.level = 0.05,
  power = NULL,
  n.events = NULL,
  alloc.ratio = 1,
  non.inf.margin = 0,
  alternative = c("two.sided", "one.sided")
)
}
\arguments{
\item{HR}{Hazard ratio of experimental group vs. control group.}

\item{sig.level}{Significance level with default 0.05.}

\item{power}{Desired power.}

\item{n.events}{Required total number of events.}

\item{alloc.ratio}{Allocation ratio: Ratio of the number of patients in the
experimental group divided by the number of patients in the control group.}

\item{non.inf.margin}{Non-inferiority margin. Default is 0 and implies a
superiority trial.}

\item{alternative}{"two.sided" (default) or "one.sided". Specifies whether
the alternative hypothesis is one- or two-sided.}
}
\value{
If \code{power} is specified, the function returns the required total
number of events. If \code{n.events} is specified, the function returns the
power.
}
\description{
For a given hazard ratio, significance level, and power,
the function returns the required total number of events. \cr
}
\details{
For a given hazard ratio, significance level, and number of events,
the function returns the corresponding power.
}
\note{
Replaces \code{\link{NumEvents}}.
}
\examples{

survEvents(HR = 0.65, sig.level = 0.05, power = 0.9, alloc.ratio = 1,
           non.inf.margin = 0, alternative = "two.sided")

survEvents(HR = 0.65, sig.level = 0.05, power = 0.9, alloc.ratio = 1,
           non.inf.margin = .5, alternative = "one.sided")

survEvents(HR = 0.65, sig.level = 0.05, n.events = 35, alloc.ratio = 1,
           non.inf.margin = 0, alternative = "two.sided")

survEvents(HR = 0.65, sig.level = 0.05, n.events = 35, alloc.ratio = 1,
           non.inf.margin = 1, alternative = "one.sided")

}
\references{
Collett, D. (2015). \emph{Modelling Survival Data in Medical
Research}.  New York: Chapman and Hall/CRC.

Schoenfeld, D.A. (1983). Sample-Size Formula for the Proportional-Hazards
Regression Model. \emph{Biometrics}, \bold{39}, 499--503.
}
\author{
Uriah Daugaard and Leonhard Held \cr \email{leonhard.held@uzh.ch}
}
