% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableOR.R
\name{tableOR}
\alias{tableOR}
\title{Results table with odds ratios from logistic regression models for binary or
ordinal data}
\usage{
tableOR(
  model,
  caption = "",
  label = "",
  size = "scriptsize",
  factorNames = NULL,
  table.placement = "ht",
  refLevels = NULL,
  lang = "english",
  short = FALSE,
  latex = TRUE,
  rmStat = FALSE,
  Wald = FALSE
)
}
\arguments{
\item{model}{Object of class \code{"polr"} or \code{"glm"} with family binomial.}

\item{caption}{Character vector of length 1 containing the table caption.}

\item{label}{Character vector of length 1 containing the reference label of
the Latex table.}

\item{size}{Character vector of length 1 setting LaTeX font site, e.g. "small",
"scriptsize", etc.}

\item{factorNames}{A character vector of size k number of factors or
regressor with custom factor labels.}

\item{table.placement}{LaTeX table positioning. The default is "ht".
See \code{?print.xtable} for more information.}

\item{refLevels}{A character vector of size k number of regressors with
custom reference level names. This argument is usuful for objects of class
\code{glm()}, which does not store the reference level labels internally.}

\item{lang}{Language of the table, either "english" (default) or "german".}

\item{short}{If \code{TRUE}, factor names are removed from factor
levels. Default is \code{FALSE}.}

\item{latex}{If \code{TRUE} (default) LateX output is produced.}

\item{rmStat}{Logical, if \code{FALSE} (default) output table inludes test
statistics.}

\item{Wald}{Logical, if \code{FALSE} (default) Wilson confidence intervals
are computed.}
}
\value{
Depending on the value of the argument \code{latex}, the function
either prints LaTeX code or returns a data frame.
}
\description{
Result table with odds ratios, 95\%-CI, test statistics, and p-values from
logistic regression models for binary or ordinal variables.
}
\examples{

dat = carData::TitanicSurvival
dat$survived = relevel(dat$survived, ref = "yes") # relevel: baseline is survived yes.
model = glm(survived ~ sex + age + passengerClass, data = dat, family = binomial())
labels = c("female", "1st") # reference levels of the two categorial variables
tableOR(model, latex = FALSE, short = TRUE, refLevels = labels, 
    	   caption = "Changes in odds for risk of death in the Titanic tragedy.")

## using log regression for ordinal data
dat$passengerClass = factor(dat$passengerClass, ordered = TRUE)
model = MASS::polr(passengerClass ~ sex + age, data = dat, Hess = TRUE)
tableOR(model, latex = FALSE, short = TRUE, 
	caption = "Changes in odds for being in a lower class, i.e. 2nd or 3rd class")

}
\seealso{
\code{\link[xtable]{xtable}}, \code{\link[xtable]{print.xtable}}
}
\author{
Simon Schwab
}
