% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibullReg.R
\name{weibullDiag}
\alias{weibullDiag}
\title{Diagnostic plot of adequacy of Weibull distribution}
\usage{
weibullDiag(formula, data, labels = NULL)
}
\arguments{
\item{formula}{A formula containing a \code{\link{Surv}} object, should only
contain one categorical predictor, or a set of indicators describing only
one predictor.}

\item{data}{Data set.}

\item{labels}{A vector containing labels for the plotted lines.}
}
\value{
Produces a plot of log Time vs. log Estimated Cumulative Hazard for
each level of the predictor (similarly to what can be obtained using
\code{\link{plot.survfit}} and the \code{fun = "cloglog"} option), as well
as a data set containing that information.
}
\description{
This function constructs a diagnostic plot of the adequacy of the Weibull
distribution for survival data with respect to one categorical covariate. If
the Weibull distribution fits the data well, then the lines produced should
be linear and parallel.
}
\details{
As discussed in Klein and Moeschberger (2003), one method for checking the
adequacy of the Weibull model with a categorical covariate is to produce
stratified Kaplan-Meier estimates (KM), which can be transformed to estimate
the log cumulative hazard for each stratum. Then in a plot of \eqn{\log(t)}
versus \eqn{\log(-\log(KM))}, the lines should be linear and parallel. This
can be seen as the log cumulative hazard for the Weibull distribution is

\deqn{\log H(t) = \log \lambda + \alpha \log t.}
}
\examples{

if (requireNamespace("prodlim")) {
    data(larynx)
    fm <- 
    weibullDiag(formula = Surv(time, death) ~ stage, data = larynx,
                labels = c("Stage I", "Stage II","Stage III", "Stage IV"))
}

}
\references{
Klein, J. and Moeschberger, M. (2003). \emph{Survival analysis:
techniques for censored and truncated data}.  Springer.
}
\seealso{
Requires packages \pkg{survival} and \pkg{prodlim}. A similar plot
can be produced using \code{\link{plot.survfit}} and the option \code{fun =
"cloglog"}.
}
\author{
Sarah R. Haile
}
\keyword{regression}
\keyword{survival}
