\name{wiedat2b}
\Rdversion{1.1}
\alias{wiedat2b}
\docType{data}
\title{Data of two pancreatic Ca biomarkers}
\description{This dataset provides two paired 
  diagnostic marker measurements for diagnosis of pancreatic cancer}
\usage{data(wiedat2b)}
\format{
  A data frame with 141 observations on the following 3 variables.
  \describe{
    \item{\code{y1}}{Biomarker 1}
    \item{\code{y2}}{Biomarker 2}
    \item{\code{d}}{Disease: 0 = non-cancer, 1 = cancer.}
  }
}
\source{This dataset is used in Wieand et al. (1989) and Pepe (2003).
}
\references{
Wieand S, Gail MH, James BR, and James KL. A family of
      nonparametric statistics for comparing diagnostic markers with paired or
      unpaired data. Biometrika 76(3):585-92. 1989.

      Pepe, M.S. (2003) \emph{The statistical evaluation of medical tests for classification and prediction}.
Oxford: Oxford University Press.
}

\keyword{datasets}
