% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blandAltman.R
\name{blandAltman}
\alias{blandAltman}
\alias{BlandAltman}
\alias{blandaltman}
\title{Bland-Altman plot to assess agreement between two continuous variables}
\usage{
blandAltman(
  x,
  y,
  conf.level = 0.95,
  group = NULL,
  xlab = "average of measurements",
  ylab = "difference of measurements",
  main = "",
  ylim = NULL,
  plot = TRUE
)
}
\arguments{
\item{x}{Measurements of first method.}

\item{y}{Measurements of second method. Must have the same length as \code{x}.}

\item{conf.level}{Confidence level for confidence intervals around the limits
of agreement. The default value is 0.95.}

\item{group}{Factor that enables different colors for points
(e.g. to group observations by observation date).}

\item{xlab}{Label for x-axis.}

\item{ylab}{Label for y-axis.}

\item{main}{Title of the plot.}

\item{ylim}{Vector of length 2 containing upper and lower limits of the y-axis.}

\item{plot}{If \code{TRUE}, the plot is generated.}
}
\value{
A list containing the following elements:
\item{difference.mean}{Mean difference.} \item{difference.sd}{Standard
deviation of differences.} \item{difference.se}{Standard error of mean
differences.} \item{upper.agreement.limit}{Upper limit of agreement.}
\item{lower.agreement.limit}{Lower limit of agreement.}
\item{agreement.limit.se}{Standard error of limits of agreement.}
\item{t.value}{\eqn{t}-quantile used for computation of limits of
agreement.}
}
\description{
Agreement between two continuous measurements is usually assessed by
plotting the difference vs. the mean of the two measurements, the so-called
Bland-Altman plot. Methods can be used interchangeably if all observations
in the Bland-Altman lie between the mean +/- two standard deviations,
provided that differences in this range were not clinically important. Note
that the function requires pairwise observations.
}
\examples{

n <- 50
meas1 <- rnorm(n, 20, 4)
meas2 <- meas1 + rnorm(n, 0.5, 1)
time <- sample(c(rep(1, n / 2), rep(2, n / 2)))

blandAltman(x = meas1, y = meas2, conf.level = 0.95,
            xlab = "mean of measurements", ylab = "difference of measurements", 
            main = "Assess agreement between measurement 1 and measurement 2")

blandAltman(x = meas1, y = meas2, conf.level = 0.95, group = time,
            xlab = "mean of measurements", ylab = "difference of measurements",
            main = "Assess agreement between measurement 1 and measurement 2")

}
\references{
Bland, J.M. and Altman, D.G. (1986). Statistical methods for
assessing agreement between two methods of clinical measurement.
\emph{Lancet}, \bold{1}, 307--310.
}
\author{
Kaspar Rufibach, \email{kaspar.rufibach@gmail.com}
}
\keyword{aplot}
\keyword{dplot}
\keyword{htest}
