% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineSubgroups.R
\name{combineSubgroups}
\alias{combineSubgroups}
\title{Combines summary data across subgroups}
\usage{
combineSubgroups(n, means, variances, treatment)
}
\arguments{
\item{n}{Vector with sample sizes in each subgroup.}

\item{means}{Vector with sample means.}

\item{variances}{Vector with sample variances.}

\item{treatment}{Indicates which entries of each vector belong to the
treatment group.}
}
\value{
Gives overall difference in means with a standard error.
}
\description{
Combines summary data for a continuous outcome given in subgroups.
}
\examples{

combineSubgroups(n = c(10, 20, 30, 40), means = c(12, 11, 10, 9),
                 variances = c(3, 4, 3, 4), treatment = c(2, 4))

}
\author{
Leonhard Held
}
