% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confIntAUC.R
\name{confIntAUCbinorm}
\alias{confIntAUCbinorm}
\alias{AUCbinorm}
\title{ROC curve and (bootstrap) confidence interval for AUC in the binormal model}
\usage{
confIntAUCbinorm(
  cases,
  controls,
  conf.level = 0.95,
  method = c("boot", "wald"),
  replicates = 1000,
  grid = 100,
  var.equal = TRUE
)
}
\arguments{
\item{cases}{Values of the continuous variable for the cases.}

\item{controls}{Values of the continuous variable for the controls.}

\item{conf.level}{Confidence level for confidence interval.}

\item{method}{Either "boot" for bootstrap CI or "wald" for a Wald CI.}

\item{replicates}{Number of boostrap replicates. Only used if
\code{method = 'boot'}.}

\item{grid}{Number of intervals to split \eqn{[0, 1]}.
Used to compute the ROC curve if \code{method = 'boot'}.}

\item{var.equal}{Logical with default TRUE.
Are the variances assumed to be equal or not?
Only used if \code{method = 'wald'}.
Currently, only \code{var.equal = TRUE} is supported.}
}
\value{
The results for \code{method = 'boot'} are
\item{a}{The values of a.}
\item{b}{The values of b.}
\item{x.val}{The values on the \eqn{x}-axis of a ROC plot.}
\item{y.val}{The values on the \eqn{y}-axis of a ROC plot, i.e. the binormal ROC curve.}
\item{auc}{Area under the ROC curve.}
\item{lowBootCI}{Lower limit of bootstrap confidence interval.}
\item{upBootCI}{Upper limit of bootstrap confidence interval.}

The results for \code{method = 'wald'} are
\item{a}{The values of a.}
\item{b}{The values of b.}
\item{auc}{Area under the ROC curve.}
\item{lowCI}{Lower limit of confidence interval.}
\item{upCI}{Upper limit of confidence interval.}
}
\description{
This function computes for values of a continuous variable of a group of
cases and a group of controls the ROC curve in the binormal model.
Additionally, the AUC including a confidence interval is provided, with
bootstrap or with Wald, assuming the variances are equal (the assumption of
unequal variances has yet to be implemented).
}
\note{
The Wald approach is documented in a vignette: see
\code{vignette("aucbinormal")}
}
\examples{

## simulate data
## --------------
set.seed(1977)
controls <- rnorm(n = 50)
cases <-  rnorm(n = 40, mean = 0.5, sd = 1.5)

## summary of ROC curve
## --------------
res <- summaryROC(cases = cases, controls = controls, conf.level = 0.95)


## alternative bootstrap CI for AUC
## --------------
resBinormBoot <- confIntAUCbinorm(cases = cases, controls = controls, 
                                  conf.level = 0.95, replicates = 1000,
                                  grid = 100)

## alternative bootstrap CI for AUC 
## --------------
resBinormWald <- confIntAUCbinorm(cases = cases, controls = controls,
                                  conf.level = 0.95, method = "wald")

## display results
## --------------
str(res)
str(resBinormBoot)
str(resBinormWald)

## plot ROC curve
## --------------
plot(x = 0, y = 0, xlim = c(0, 1), ylim = c(0, 1), type = "l", 
     xlab = "1 - specificity", ylab = "sensitivity", pty = "s")
segments(x0 = 0, y0 = 0, x1 = 1, y1 = 1, lty = 2)
lines(x = res$x.val, y = res$y.val, type = 'l', col = 2, lwd = 2, lty = 2)
lines(x = resBinormBoot$x.val, y = resBinormBoot$y.val, type = "l",
      col = 4, lwd = 2, lty = 2)

}
\references{
Pepe, M.S. (2003) \emph{The statistical evaluation of medical
tests for classification and prediction}. Oxford: Oxford University Press.
}
\seealso{
\code{\link{summaryROC}}
}
\author{
Kaspar Rufibach (\code{method = 'boot'}) and Leonhard Held
(\code{method = 'wald'})
}
\keyword{htest}
