% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confIntCorrelation.R
\name{confIntFisherTrafo}
\alias{confIntFisherTrafo}
\title{Deprecated: use confIntCorrelation instead}
\usage{
confIntFisherTrafo(
  var1,
  var2,
  pp = c(0.025, 0.975),
  meth = "spearman",
  type = "t"
)
}
\arguments{
\item{var1}{Vector containing first variable.}

\item{var2}{Vector containing first variable.}

\item{pp}{Vector in \eqn{R^2} that contains \eqn{\alpha / 2} and \eqn{1 - \alpha/2},
where \eqn{alpha} is the confidence level of the confidence interval.}

\item{meth}{Correlation coefficient to be used: \code{pearson} or \code{spearman}.}

\item{type}{Quantile to be used: \code{z} or \code{t}.}
}
\value{
List with entries:
\item{estimate}{Value of correlation coefficient.}
\item{ci}{Computed confidence interval.}
\item{p.value}{\eqn{p}-value for a test on \eqn{\rho = 0} based on the transformation.}
\item{n}{Number of observations.}
\item{p2}{\eqn{p}-value based on the \code{R} function \code{cor.est}.}
}
\description{
Deprecated: use confIntCorrelation instead
}
