% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confIntKappa.R
\name{confIntKappa}
\alias{confIntKappa}
\title{Confidence intervals for weighted kappa and m >= 2 raters}
\usage{
confIntKappa(
  data,
  type = c("not Cohen", "Cohen"),
  weights = c("squared", "absolute"),
  m = 1000,
  conf.level = 0.95
)
}
\arguments{
\item{data}{m x n matrix or data.frame containing data from m subjects and n raters.}

\item{type}{String defining the type of confidence interval.
If not equal to "not Cohen" (default), the weighted
version for ordered ratings is computed.
If "Cohen", Cohen's unweighted \eqn{kappa} is computed, i.e.
ratings are assumed to be nominal.}

\item{weights}{Define weights to be used if ordered ratings are compared.
Can be "squared" (default) or "absolute".
Only used if \code{type != "Cohen"}.}

\item{m}{Number of bootstrap samples to be generated.}

\item{conf.level}{Confidence level for confidence interval. Default is 0.95.}
}
\value{
A list containing: \item{n}{Number of observations used to compute
confidence intervals.} \item{kappa}{Computed \code{kappa}.}
\item{boot.quant}{Confidence interval based on quantiles of the bootstrap
distribution.}
}
\description{
Compute confidence intervals for the coefficient of agreement for two
nominal or ordered variables and two or more raters.
}
\details{
This function computes bootstrap confidence intervals for an unweighted or
weight \eqn{kappa} coefficient, based on all pairwise complete observations
in \code{data}.

Fleiss' kappa for m raters according to Conger (1980), allows
for weighting also in the case of m > 2 raters computed using the function
\code{\link[psy]{lkappa}} in package \pkg{psy}. 
For bootstrap confidence interval the package \pkg{boot} is used.
}
\examples{

## example is similar to example in ?lkappa
data("expsy", package = "psy")
set.seed(14)
confIntKappa(data = expsy[, c(11,13,15)], type = "not Cohen", weights = "absolute",
             m = 200, conf.level = 0.95)

}
\references{
Conger, A.J. (1980), \emph{Integration and generalisation of
Kappas for multiple raters}, Psychological Bulletin, 88, 322-328.
}
\seealso{
\code{\link[psy]{lkappa}}, \code{\link[boot]{boot}}
}
\keyword{htest}
