% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confIntRiskDiff.R
\name{confIntOddsRatio}
\alias{confIntOddsRatio}
\title{Confidence interval for an odds ratio}
\usage{
confIntOddsRatio(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{Vector of length 2, number of successes in each group.}

\item{n}{Vector of length 2, total number of trials in each group.}

\item{conf.level}{Confidence level for confidence interval. Default value is 0.95.}
}
\value{
A vector containing the odds ratio and the limits of the confidence interval.
}
\description{
Provides a confidence interval for an odds ratio.
The method is based on a Wald interval for the log odds ratio.
Used by \code{\link{confIntDiagnostic}}.
}
\examples{
x <- c(30, 50)
n <- c(100, 120)
confIntOddsRatio(x, n)
}
\references{
DG Altman, D Machin, TN Bryant, MJ Gardner.
\emph{Statistics with confidence}, 2nd Edition, 2000, Chapter 7
}
\seealso{
\code{\link{confIntDiagnostic}}, \code{\link{confIntRiskRatio}}, \code{\link{confIntRiskDiff}}.
}
\author{
Leonhard Held
}
\keyword{htest}
