% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confIntAUC.R
\name{confIntPairedAUCDiff}
\alias{confIntPairedAUCDiff}
\title{Confidence interval for the difference in AUC based on two paired samples}
\usage{
confIntPairedAUCDiff(cases, controls, conf.level = 0.95)
}
\arguments{
\item{cases}{Matrix with values of the continuous variable for the cases.
First column gives values of test A, second gives values of test B.}

\item{controls}{Matrix with values of the continuous variable for the
controls. First column gives values of test A, second gives values of test
B.}

\item{conf.level}{Confidence level for confidence interval. Default is 0.95.}
}
\value{
A data.frame with the estimates and confidence limits for the AUC of the two
tests and their difference.
}
\description{
Computes confidence interval for the difference in the area under the curve
based on two paired samples.
}
\examples{

data(wiedat2b)
ind <- wiedat2b[,3]
cases <- wiedat2b[ind == 1, 1:2]
controls <- wiedat2b[ind == 0, 1:2]
confIntPairedAUCDiff(cases = cases, controls = controls)

}
\references{
Pepe, M.S. (2003) \emph{The statistical evaluation of medical
tests for classification and prediction}. Oxford University Press.
}
\seealso{
\code{\link{confIntAUC}}, \code{\link{confIntIndependentAUCDiff}},
\code{\link{standardErrorAUCDiff}}
}
\author{
Leonhard Held
}
