% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confIntRate.R
\name{confIntRate}
\alias{confIntRate}
\alias{waldRate}
\alias{wilsonRate}
\alias{confIntRateWald}
\alias{confIntRateWilson}
\title{Wald and Wilson confidence intervals for Poisson rates}
\usage{
confIntRate(x, t, conf.level = 0.95)

confIntRateWald(x, t, conf.level = 0.95)

confIntRateWilson(x, t, conf.level = 0.95)
}
\arguments{
\item{x}{Number of events.}

\item{t}{Total observation time.}

\item{conf.level}{Confidence level for confidence interval.}
}
\value{
\code{confIntRate} returns a data.frame with columns: \code{x}, \code{t}, \code{type},
\code{rate}, \code{lower}, \code{upper}.

\code{confIntRateWald}returns a matrix with columns
\code{rate}, \code{lower}, and \code{upper}.

\code{confIntRateWilson} returns a matrix with columns
\code{rate}, \code{lower}, and \code{upper}.
}
\description{
Compute a confidence interval for Poisson rates using the Wald and Wilson methods.
}
\examples{

confIntRateWald(x = 1, t = 2.2)
confIntRateWilson(x = 1, t = 2.2)
confIntRate(x = 3:1, t = seq(from = 1, to = 2.2, length.out = 3))

}
\references{
Held, L., Rufibach, K. and Seifert, B. (2013). Medizinische
Statistik - Konzepte, Methoden, Anwendungen. Section 8.2.
}
\seealso{
\code{\link{confIntRateDiff}}, package \pkg{Hmisc}.
}
\author{
Leonhard Held
}
