% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confIntRate.R
\name{confIntRateDiff}
\alias{confIntRateDiff}
\title{Wald and Wilson confidence intervals for Poisson rates differences}
\usage{
confIntRateDiff(x, t, conf.level = 0.95)
}
\arguments{
\item{x}{vector of length 2, number of events in each group.}

\item{t}{vector of length 2, total observation time in each group.}

\item{conf.level}{Confidence level for confidence interval.}
}
\value{
A list with the entries:
\item{rd}{Estimated rate difference.}
\item{CIs}{Data.frame containing confidence intervals for the rate difference.}
}
\description{
Compute a confidence interval for Poisson rate differences using the
Wald and Wilson methods.
}
\examples{

confIntRateDiff(x = c(30, 50), t = c(100, 120.1))

}
\references{
Held, L., Rufibach, K. and Seifert, B. (2013). Medizinische
Statistik - Konzepte, Methoden, Anwendungen. Section 8.2.
}
\seealso{
\code{\link{confIntRateWilson}}, \code{\link{confIntRateWald}}
}
\author{
Leonhard Held
}
\keyword{htest}
