% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confIntRiskDiff.R
\name{confIntRiskDiff}
\alias{confIntRiskDiff}
\title{Confidence interval for a risk differences,}
\usage{
confIntRiskDiff(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{Vector of length 2, number of successes in each group.}

\item{n}{Vector of length 2, total number of trials in each group.}

\item{conf.level}{Confidence level for confidence interval. Default value is
0.95.}
}
\value{
A list with the entries: \itemize{ \itemrdEstimated risk difference.
\itemCIsDataframe containing confidence intervals for the risk difference. }
}
\description{
Two methods to compute a confidence interval for a risk difference based on
Wald and Wilson confidence intervals for the individual risks are provided.
}
\examples{

x <- c(30, 50)
n <- c(100, 120)
confIntRiskDiff(x, n)$CIs

}
\references{
DG Altman, D Machin, TN Bryant, MJ Gardner. \emph{Statistics
with confidence}, 2nd Edition, 2000, Chapter 6
}
\seealso{
\code{\link{wilson}}, \code{\link{confIntRiskRatio}},
\code{\link{confIntOddsRatio}}.
}
\author{
Leonhard Held
}
\keyword{htest}
