% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confIntDiagnostic.R
\name{confIntRiskRatio}
\alias{confIntRiskRatio}
\title{Confidence interval for a risk ratio}
\usage{
confIntRiskRatio(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{Vector of length 2, number of successes in each group.}

\item{n}{Vector of length 2, total number of trials in each group.}

\item{conf.level}{Confidence level for confidence interval. Default value is
0.95.}
}
\value{
A vector containing the risk ratio and the limits of the confidence
interval.
}
\description{
Provides a confidence interval for a risk ratio. The method is based on the
Wald interval for the log risk ratio.
}
\examples{

x <- c(30, 50)
n <- c(100, 120)
confIntRiskRatio(x = x, n = n)

}
\references{
DG Altman, D Machin, TN Bryant, MJ Gardner. \emph{Statistics
with confidence}, 2nd Edition, 2000, Chapter 7
}
\seealso{
\code{\link{confIntDiagnostic}}, \code{\link{confIntRiskRatio}},
\code{\link{confIntOddsRatio}}.
}
\author{
Leonhard Held
}
\keyword{htest}
