% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faganPlot.R
\name{faganPlot}
\alias{faganPlot}
\alias{faganLine}
\title{Fagan-Nomogram}
\usage{
faganPlot(
  probPreInit = c(0.1, 0.2, 0.5, 1, 2, 5, 10, 20, 30, 40, 50, 70, 85),
  cex = 0.7,
  language = c("english", "german"),
  title = "Fagan - Nomogramm"
)

faganLine(probPre, likRatio = c(1, 1))
}
\arguments{
\item{probPreInit}{Pre-test probabilities in percent to be used for the vertical line
to the left. A number or a vector with entries in \eqn{(0,100)}.}

\item{cex}{Text size.}

\item{language}{Either "english" (default) or "german".}

\item{title}{Title of the plot.}

\item{probPre}{Numric vector specifying the pre-test probability.}

\item{likRatio}{Numeric vector specifying the likelihood ratio.}
}
\value{
\code{faganPlot} draws a Fagan-Nomogram.

\code{faganLine} adds lines to a plot obtained by \code{faganPlot}.
}
\description{
Draw the Fagan-Nomogram.
}
\examples{

# empty Fagan plot
faganPlot()

# add Fagan lines
faganLine(probPre = 0.75, likRatio = c(0.9 / 0.37, 0.1 / 0.63))

}
\references{
Fagan, T.J. (1975). Letter: Nomogram for Bayes Theorem. \emph{N.
Engl. J. Med}, \bold{293}, 257.
}
\author{
Kaspar Rufibach \cr \email{kaspar.rufibach@gmail.com}
}
\keyword{aplot}
\keyword{dplot}
