% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioStatUZH-package.R
\docType{data}
\name{fischdaten}
\alias{fischdaten}
\title{Data of the 'Fischstudie'}
\format{
A data frame with 308 observations on the following 16 variables.
\describe{ \item{list("patnr")}{Patient number.} \item{list("geschl")}{Sex:
0 = female, 1 = male.} \item{list("alter")}{Age in years.}
\item{list("groesse")}{Height in cm.} \item{list("gew0")}{Weight at study
entry, in kg.} \item{list("gew15")}{Weight at 15 days, in kg.}
\item{list("gew28")}{Weight at 28 days, in kg.}
\item{list("chol0")}{Cholesterin level at study entry, in mg/dl.}
\item{list("chol15")}{Cholesterin level at 15 days, in mg/dl.}
\item{list("chol28")}{Cholesterin level at 28 days, in mg/dl.}
\item{list("fisch")}{Received fish (1) or not (0).}
\item{list("kalorie")}{Received reduced food (1) or not (0).} }
}
\source{
This dataset is used in Schumacher and Schulgen (2002) and Held et
al. (2013).
}
\description{
This dataset provides data of 308 patients enrolled in the 'Fischstudie',
described in Schumacher and Schulgen (2002).
}
\references{
Held, Leonhard, Rufibach, Kaspar, and Seifert, Burkhardt (2013).
Medizinische Statistik. Konzepte, Methoden, Anwendungen. Pearson Verlag,
Halbergmoos, Germany.

Schumacher, Martin and Schulgen, Gabi. (2002). Methodik klinischer Studien.
Springer, Berlin, Germany.
}
\keyword{datasets}
