\name{fischdaten}
\Rdversion{1.1}
\alias{fischdaten}
\docType{data}
\title{Data of the 'Fischstudie'}
\description{This dataset provides data of 308 patients enrolled in the 'Fischstudie', described in Schumacher and 
Schulgen (2002).}
\usage{data(fischdaten)}
\format{
  A data frame with 308 observations on the following 16 variables.
  \describe{
    \item{\code{patnr}}{Patient number.}
    \item{\code{geschl}}{Sex: 0 = female, 1 = male.}
    \item{\code{alter}}{Age in years.}
    \item{\code{groesse}}{Height in cm.}
    \item{\code{gew0}}{Weight at study entry, in kg.}
    \item{\code{gew15}}{Weight at 15 days, in kg.}
    \item{\code{gew28}}{Weight at 28 days, in kg.}
    \item{\code{chol0}}{Cholesterin level at study entry, in mg/dl.}
    \item{\code{chol15}}{Cholesterin level at 15 days, in mg/dl.}
    \item{\code{chol28}}{Cholesterin level at 28 days, in mg/dl.}
    \item{\code{fisch}}{Received fish (1) or not (0).}
    \item{\code{kalorie}}{Received reduced food (1) or not (0).}
  }
}
\source{This dataset is used in Schumacher and Schulgen (2002) and was adapted in the lecture notes 
Held et al. (2008).
}
\references{
Held, Leo, Rufibach, Kaspar, and Seifert, Burkhardt (2008).
Einfuehrung in die Biostatistik.
Lecture Notes, Biostatistics unit, Institute of Social and Preventive Medicine, University of Zurich.

Schumacher, Martin and Schulgen, Gabi. (2002).
Methodik klinischer Studien.
Springer, Berlin.
}
\keyword{datasets}
