% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatPercent.R
\name{formatPercent}
\alias{formatPercent}
\title{Format a Numeric Proportion}
\usage{
formatPercent(x, digits = 1)
}
\arguments{
\item{x}{numeric vector of proportions}

\item{digits}{number of digits}
}
\description{
Takes a number and formats it as a percentage.  Used by
\code{\link{confIntIndependentProportion}}.
}
\examples{

formatPercent(c(0.115, 0.5))  # "11.5\%" "50.0\%"

}
\author{
Leonhard Held
}
