% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationSamplePlot.R
\name{populationSamplePlot}
\alias{populationSamplePlot}
\title{Plot of population and sample relationship}
\usage{
populationSamplePlot(language = c("german", "english"))
}
\arguments{
\item{language}{Language of the text of the plot.}
}
\value{
Generates a plot.
}
\description{
Plot visualizing the connection between a population
and a sample drawn from it.
}
\examples{
populationSamplePlot()
}
\author{
Kaspar Rufibach \cr \email{kaspar.rufibach@gmail.com}
}
\keyword{aplot}
\keyword{dplot}
