\name{populationSamplePlot}
\alias{populationSamplePlot}
\title{Plot that shows population and sample drawn from it}
\description{We frequently use this plot in presentations and teaching, to explain the connection between
a population and a sample drawn from it.}
\usage{populationSamplePlot(language = c("german", "english")[1])}
\arguments{\item{language}{Language of the text pieces added to the plot.}}
\value{Nothing is ouutput, only a plot is generated,}

\author{Kaspar Rufibach \cr \email{kaspar.rufibach@gmail.com}}

\keyword{dplot}
\keyword{aplot}
