% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileCumInc.R
\name{quantileCumInc}
\alias{quantileCumInc}
\title{Quantile of a cumulative incidence estimate}
\usage{
quantileCumInc(time, event, group, quant = 0.5)
}
\arguments{
\item{time}{Event times, censored or observed.}

\item{event}{Event indicator.}

\item{group}{Quantiles can be computed for several survival curves, defined
by \code{group}.}

\item{quant}{Quantile to be computed. Real number in \eqn{[0, 1]}.}
}
\value{
data.frame with quantiles estimates for each event-group combination.
}
\description{
In survival analysis, often some quantiles (e.g., the median) of an
estimated cumulative incidence function are of interest. This function
computes quantiles of cumulative incidence functions using
\code{cuminc} of the package \pkg{cmprsk}.
}
\examples{

if (require("cmprsk")) {
  ## illustrate on simulated data
  set.seed(1977)
  n <- 180
  time <- rgamma(n, 2, 1)
  status <- factor(sample(rep(0:2, n / 3)), levels = 0:2,
                   labels = c("censored", "Time to 2nd tumor", "Death"))
  x <- factor(round(runif(n, 1, 2)), levels = 1:2,
              labels = c("Tmt A", "Tmt B"))
  
  ## plot
  plot(cuminc(time, status, x))

  ## median time to event
  quantileCumInc(time, status, group = x, quant = 0.25)
}

}
\seealso{
\code{cuminc} and \code{timepoint} in the package \pkg{cmprsk}.
}
\author{
Kaspar Rufibach \cr \email{kaspar.rufibach@gmail.com}
}
\keyword{htest}
\keyword{survival}
