% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSizeWilcoxTwoSample.R
\name{sampleSizeWilcoxTwoSample}
\alias{sampleSizeWilcoxTwoSample}
\title{Sample size for two sample Wilcoxon (Mann-Whitney) test}
\usage{
sampleSizeWilcoxTwoSample(
  delta = 0.75,
  sig.level = 0.05,
  power = 0.9,
  ratio = 0.5,
  two.sided = TRUE
)
}
\arguments{
\item{delta}{Probability \eqn{P(Y_i > X_j)}.}

\item{sig.level}{Significance level with default 0.05.}

\item{power}{Power with default 0.9.}

\item{ratio}{Ratio of the sample size of group 1 to the total samples size:
\code{ratio} \eqn{= m / (m + n)}, where \eqn{m} and \eqn{m} are the sample
sizes of group 1 and 2, respectively.
A \code{ratio} of 0.5 implies equal group sizes.}

\item{two.sided}{If \code{TRUE}, the test is two-sided.
Otherwise, it is one-sided.}
}
\value{
data.frame with columns: 
\item{delta}{\code{delta} from input.}
\item{sig.level}{\code{sig.level} from input.}
\item{power}{\code{power} from input.}
\item{ratio}{\code{ratio} from input.}
\item{two.sided}{\code{two.sided} from input.}
\item{m}{Sample size of the group 1.}
\item{n}{Sample size of the group 2.}
}
\description{
Compute the sample size required to test the hypothesis that two sets of independent samples,
\eqn{X_1, ..., X_m} and \eqn{Y_1, ..., Y_n} come from the
same population against that \eqn{Y}'s is to be larger than \eqn{X}'s.
}
\examples{

sampleSizeWilcoxTwoSample(delta = 0.75)
sampleSizeWilcoxTwoSample(delta = 0.85)
sampleSizeWilcoxTwoSample(delta = 0.85, power = 0.8)
sampleSizeWilcoxTwoSample(delta = 0.85, sig.level = 0.01)
sampleSizeWilcoxTwoSample(delta = 0.85, two.sided = FALSE)

tmp <- lapply(X = seq(0.55, 0.95, .1),
              FUN = function(delta) sampleSizeWilcoxTwoSample(delta = delta, ratio = .25))
do.call(what = rbind, args = tmp)

}
\references{
Nother, G.E. (1987). Sample Size Determination for Some Common
Nonparametric Tests \emph{JASA}, \bold{82}, 644--647.
}
\author{
Kaspar Rufibach
}
\keyword{htest}
