% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confIntAUC.R
\name{standardErrorAUCDiff}
\alias{standardErrorAUCDiff}
\title{Standard Error of a AUC difference}
\usage{
standardErrorAUCDiff(cases, controls)
}
\arguments{
\item{cases}{Matrix with values of the continuous variable for the cases.
First column gives values of test A, second gives values of test B.}

\item{controls}{Matrix with values of the continuous variable for the
controls. First column gives values of test A, second gives values of test B.}
}
\value{
The standard error.
}
\description{
Computes the standard error of the difference in area under the curve for
paired samples.
}
\references{
Pepe, M.S. (2003) \emph{The statistical evaluation of medical
tests for classification and prediction}. Oxford University Press.
}
\seealso{
\code{\link{confIntAUC}}, \code{\link{confIntIndependentAUCDiff}},
\code{\link{confIntPairedAUCDiff}}
}
\author{
Leonhard Held
}
\keyword{htest}
\keyword{univar}
