% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{wiedat2b}
\alias{wiedat2b}
\title{Data of two pancreatic Ca biomarkers}
\format{
A data.frame with 141 observations on the 3 variables:
\describe{
\item{\code{y1}}{Biomarker 1.}
\item{\code{y2}}{Biomarker 2.}
\item{\code{d}}{Disease: 0 = non-cancer, 1 = cancer.}
}
}
\usage{
data(wiedat2b)
}
\description{
This dataset provides two paired diagnostic marker measurements
for diagnosis of pancreatic cancer.
}
\references{
Wieand S, Gail MH, James BR, and James KL. \emph{A family of
nonparametric statistics for comparing diagnostic markers with
paired or unpaired data.} Biometrika 76(3):585-92. 1989.

Pepe, M.S. (2003) \emph{The statistical evaluation of medical tests for
classification and prediction}. Oxford University Press.
}
\keyword{data}
